First commit of snapcommunicator in its own project.
The snapcommunicator requires the flags which are found in the sitter which
needs the eventdispatcher, so dependency wise, we need the snapcommunicator
outside of the eventdispatcher. Also it is a huge project and it makes more
sense to have it as a separate item.

This will also force us to have the fluid-settings as a separate project
(soonish).