The snapcommunicator is starting to work.
I got the snapmessage tool to send messages and get replies (HELP and REGISTER).
The snapmessage tool now uses the new syntax for the list of options.
Started work on "connection lost" functionality in sendmessage. It still doesn't work right.
Documented the "sc:", "scs:", and "scu:" protocols.
Implemented those protocols inside the sendmessage. We need to move it to the snapcommunicator library.
Actually setup the dispatcher in the server.
Removed the "f_logrotate", we already have that functionality on the UDP port.
Made the f_interrupt a weak pointer so Ctrl-C works as expected.
Made all the ports use "sc::<name>_PORT" names.
Fixed the base_connection pointer use, especially the send_message() fisco and user_data<>().