We already had a raise-flag tool in the snapwebsites library. I now use that implementation as the saveflag (renamed, etc.).
Made the raise-flag output JSON or plain text when listing existing raised flags.
Added many shortnames for the commands and options of raised-flag.
Hide the logger banner.
Include a configuration file with the location of the service files for communicatord to read.
Fixed the documentation, it is communicatord (the 'd' was missing).
Also the library is under communicatord (without the /src).