Tested all listeners would properly open.
As a result, I finally found the bug in the libssl, one of the functions would burry a bind() error. I have kludge in the eventdispatcher for this issue and it works appropriately each time now.
Added a block_ip() function which blocks IPs after 3 invalid login attempts within 15 min.
Fixed ports in a few place that used the wrong one (or were not yet switch to using the constexpr).
Enhanced script used to start communicatord as a developer.
Changed "plain remote" connection to private network, not ANY, and if not the correct IP, it's a fatal error and it is not required (can be left undefined for a local instance--i.e. one computer RPC).
Added a username/password for the secure connection (without such, anyone could still use that port!)
Use SNAP_LOG_CONFIGURATION to list all the listeners we successfully establish on startup.
Switched from SECURE to ALWAYS-SECURE.
In the f_remote_ip/f_secure_ip I now put the IP and the Port.
Fixed the default URIs to use the cd:/cds: protocol.
Started work on the communicatord client helper class.
Documentation enhancements.