Avoid several crashes in the communicator daemon.
When testing whether a server name was specified on a message, make sure that the connection is already UP.
When testing the "conn" & "unix", we would always use the "conn" even if nullptr.
The msg_servicestatus() would make use o the "service" parameter without testing whether the parameter was defined.
Fixed the "message" tool: added documentation (/msg_help), fixed the supported protocols.