Fixed the raise-flag implementation:
The tool uses the library and has the ability to change it's effective user ID so it can read/write files to `/var/log/communicatord/flags`.
The library attempts a change, if it fails, it uses the tool.
Enhance the library so it can read any parameter from `flags.conf`.
Actually create the `/var/log/communicatord/flags` directory.

Added CppProcess as a dependency.