Added the stable_clock check.
This allows us to run processes which verify that the clock is stable before
getting started. This will work for any system that connects to the
communicator daemon. Once the clock is stable, we send a CLOCK_STABLE message.
If the clock never becomes stable or the client requests the CLOCK_STATUS,
then we end up sending CLOCK_UNSTABLE. Both messages have a parameter which
better defines how stable (NTP not install or NTP replied positively)
or unstable (NTP replied and the ntp-wait tool timed out or the ntp-wait
command line tool is not yet done working).