The ntp_wait cannot be given a shared pointer.
In the constructor of the stable_clock, the ntp_wait is not ready to generate
a shared pointer. So I use a bare pointer since we have a parent/child
relationship anyway.

Also fixed the --debug option of the start-for-dev.sh script. However, the
"break on exception" flag is not working right.