Make the GOSSIP more resilient.
This version keeps the gossip_connection's around. They get disabled when
a connection happens (CONNECT). And they get re-enabled if we lose the
connectionr. The further GOSSIP messages should not be necessary, unless
some of the communicator daemons cannot save the list of neighbors (improper
initialization, disk full, etc.) The advantage of having the communicator
daemons still communicating is that the one with issues can then report
those issues to the others.