Many additional fixes.
Set proper defaults (i.e. "now") to f_date & f_modified on construction of the flag.
The to_string() now show whether flag is up or down.
The to_string() always uses the "last modified" date.
If state is down, try deleting early and if it works, be done with it.
Don't temper with the gossip_connection enable flag. This is a permanent connection and the flag is handled by the eventdispatcher. Instead delete the connection and recreate it as required.
Fixed the next_wait(), we needed to use std::min().
Use the REMOTE_PORT as the default port for remote communicators. It makes more sense, I think.
Use "remote_addr" for all the remote_communicators functions (instead of just address).
Actually added the `--max-gossip-timeout <duration>` option so we can now tweak it.
Made the `--max-gossip-timeout` a duration so you can use 0.5 seconds or 500ms, etc.
Fixed the service_connection IP address, we were reading the local socket name instead of the remote peer socket address/port.