Properly protect the secure_listen=... variable.
Changed the parsing of the secure_listen=... to use the edhttp::uri instead
of parsing by hand. This includes a scheme now (cds://...).

Made notes about the various other possible changes. Especially, parsing all
the other listen=... using the URI parser.

Added a few new names.an entries for the schemes supported in the library.