Renamed set_commands() as add_commands().
That function does not replace the list of commands. It has always been
cummulative, which is what we want now. If a service dynamically adds
support for commands, then those commands can be sent to the communicator
daemon and be added to the list.

This happens when we request for a cluster lock with the cluck library.
All of a sudden a client of the cluck daemon learns how to support the
LOCKED, LOCK_FAILED, etc. commands. These need to be added or the client
will not receive them (because the communicator will prevent the forwarding).