Added a note about the shared pointer bug.
The children should just use a bare pointer back to the server and not
a shared pointer. Right now that prevents the system from quitting cleanly.
That being said, it may be that I needed the shared pointer in a few
places. If that's the case, I can at least use weak pointers instead.