Added command line option for timedate-wait
The timedate-wait tool is not installed under /usr/bin in the development
environment so I added an option to allow for the user to specify the
location.

The stable_clock class was also updated to not use the server shared pointer.
A bare pointer is sufficient and it will not prevent the destruction of the
server on an uncaught exception.