Starting a big change to support plugins.
I have an issue where I want functionality about the firewall which depends
on Prinbee but Prinbee needs to know whether the firewall is up. To support
that without having an invalid dependency, I think I can make it work
properly using plugins. So the iplock will have a plugin which checks the
status of the firewall and make it available through the communicator daemon.