Renamed the project "communicator" without the "d".
This is less confusing in regard to the library which should be called
"communicator" and the daemon "communicatord". However, because we have
a communicator in the eventdispatcher and our messengers derive from the
communicator::communicator object, C++ gets a bit confused in a few places.
However, that allows us to make two packages: one for the tools and library
and one for the daemon. This means you don't absolutely have to install
the communicator on all machines (although local services can still connect
to a communicator, it may not work as expected).

Moved the clock stability check and the loadavg functionality to separate
plugins. This will also allow us to add a firewall plugin in the iplock
project which then prinbee can benefit from.

Replaced all the children pointer_t back to the communicator daemon with
a bare pointer. This means they will not hold the daemon hostage on an
exception which does not otherwise properly remove all of these from the
parent (although I did not yet test that, it is likely that the
ed::communicator will prevent the quit, still).

Reorganized some of the README.md documentation. There are things "TODO"
in there. I will have to re-work that one day.

Fixed up various docs as I found some mistakes or old references.

Since I changed the name of the project, I also changed the name of the
user to just "communicator". This matches our other services that do not
add a "d" at the end of the user & group names.