Renamed the IPWALL commands IPLOCK because the IPWALL service is not the one we send the messages to.
Added a variable in cmake to know where to install the plugins.
Added the firewall_status parameter name.
Added the firewall active value.
Install the necessary headers for plugin development.
Commented out the timedate-wait-command option and created a .ini instead.
Added the options file directory to the environment.
Added libboost-dev as a dependency since we now use signals in our plugins.